Move some style macros into their own module, used first
